Set timezone settings to Asia/Kuala_Lumpur.
It turned out that Jekyll look at datetime when building the site.
So, if for example you has post named as:-

_posts/2016-03-14-some-post.md

And the current date is 2016-03-13, then the post will not been build
as the date is in the future than current date. Jekyll by default use
OS local time if `timezone` setting not specified. This most of the
time not an issue.

In our case however, the site is build on GitLab CI server, which using
UTC time zone. So there's a case when I push new article, the article's
date which using local time (UTC+8) is in the future than the CI server
local time, so it's not build.